Db.php: replace Tools::displayError() with a simple echo.
The problem was that Tools::displayError() does nothing. So the
error message just disappears into the void and the program ends
without any message or error flag. With the echo there will at
least be a message so that we know what went wrong.

Found and fixed by @musicpanda and discussed in issue #374.