AjaxProcessor.php: fix table restoration query.
The main reason is that the former function adds a prefix to the
table name. So you get an error that it doesn't know the table
tb_tb_access. Besides this the change also seemed to solve an
eternal loop problem but I am not sure about that.

Found and fixed by @musicpanda and discussed in issue #374.