AjaxProcessor: make file removal happen before file additions.
Before, only the comment said that it should be done this way, but
using array_merge() actually ended up doing the opposite.
Evaluated and tested with a 1.0.5 -> 1.0.7 update, which removes
files in vendor/paragonie/random_compat/ with the 1.0.5 -> 1.0.6
update, then adds them again with the 1.0.6 -> 1.0.7 one. Before,
this ended up with these files missing.

Whether array_merge() always places elements of the second array
in front, or whether it places the larger of two given arrays in
front, was not tested. Documentation doesn't explicitely state an
expected order of the resulting array.

Related to issue #619 in thirty bees core.